List the invoked functions of a class
